refactoring: clean up ide warnings and migrate away from obsolete methods
