+++ src/Umbraco.Commerce.ProductFeeds.Web/Umbraco.Commerce.ProductFeeds.Web.csproj
@@ -4,7 +4,7 @@
     <Title>Umbraco Commerce Product Feeds Web</Title>
     <Description>Contains apis</Description>
     <Nullable>enable</Nullable>
-    <NoWarn>IDE0007;CA2007</NoWarn>
+    <NoWarn>IDE0007;IDE0058;CA2007;IDE0022;IDE0021;IDE0290;IDE0063;</NoWarn>
     <StaticWebAssetBasePath>App_Plugins/umbracocommerceproductfeeds</StaticWebAssetBasePath>
   </PropertyGroup>
